avocado.core.tree: Move mux-related TreeNodes to yaml plugin
The basic `avocado.core.tree` implementation should not cover anything
related to multiplexer and multiplexation, it should only provide the
basic interface to arrange key/value pairs into tree-like structures.
This is only a first part, which moves the TreeNode structure away, more
work on that area follows.
